Revert "mupen64: allocate 1MiB on heap instead of stack"
This reverts commit 2c75e376bf2ed0d21cb7c02ee33ab30676fa2ee7.

+++ mupen64plus-core/src/main/main.c
@@ -835,7 +835,7 @@ extern audio_plugin_functions dummy_audio;
 
 unsigned int r4300_emumode;
 
-unsigned long rdram_size;
+uint32_t rdram_size;
 struct file_storage eep;
 struct file_storage fla;
 struct file_storage sra;
